+++ examples/web-app/src/index.ts
@@ -10,14 +10,22 @@ const system = createSystem([[postgres, redis], app, httpServer], {
 });
 
 system.on(SystemEvent.StartInitiated, ({ name }) => console.log(`${name} system is starting`));
-system.on(SystemEvent.StartSucceeded, ({ name }) => console.log(`${name} system started`));
+system.on(SystemEvent.StartSucceeded, ({ name, duration }) =>
+  console.log(`${name} system started in ${duration.toFixed(0)}ms`),
+);
 system.on(SystemEvent.StartFailed, ({ name, error }) => console.error(`${name} system failed to start`, error));
 system.on(SystemEvent.StopInitiated, ({ name }) => console.log(`${name} system is stopping`));
-system.on(SystemEvent.StopSucceeded, ({ name }) => console.log(`${name} system stopped`));
+system.on(SystemEvent.StopSucceeded, ({ name, duration }) =>
+  console.log(`${name} system stopped in ${duration.toFixed(0)}ms`),
+);
 system.on(SystemEvent.StopFailed, ({ name, error }) => console.error(`${name} system failed to stop`, error));
 
-system.on(ComponentEvent.StartSucceeded, ({ name }) => console.log(`${name} component started`));
-system.on(ComponentEvent.StopSucceeded, ({ name }) => console.log(`${name} component stopped`));
+system.on(ComponentEvent.StartSucceeded, ({ name, duration }) =>
+  console.log(`${name} component started in ${duration.toFixed(0)}ms`),
+);
+system.on(ComponentEvent.StopSucceeded, ({ name, duration }) =>
+  console.log(`${name} component stopped in ${duration.toFixed(0)}ms`),
+);
 system.on(ComponentEvent.StartFailed, ({ name, error }) => console.error(`${name} component failed to start`, error));
 system.on(ComponentEvent.StopFailed, ({ name, error }) => console.error(`${name} component failed to stop`, error));